+++ app/src/androidTest/java/com/mapbox/mapboxsdk/plugins/locationlayer/LocationLayerPluginTest.java
@@ -14,7 +14,6 @@ import com.mapbox.mapboxsdk.utils.OnMapReadyIdlingResource;
 
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -25,14 +24,18 @@ import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
+import static com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerConstants.ACCURACY_LAYER;
+import static com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerConstants.BACKGROUND_LAYER;
+import static com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerConstants.BEARING_LAYER;
+import static com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerConstants.FOREGROUND_LAYER;
+import static com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerConstants.LOCATION_SOURCE;
+import static com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerConstants.NAVIGATION_LAYER;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 @RunWith(AndroidJUnit4.class)
 @SuppressWarnings( {"MissingPermission"})
 public class LocationLayerPluginTest {
-
-  private static final double DELTA = 1E-10;
 
   @Rule
   public ActivityTestRule<LocationLayerModesActivity> rule = new ActivityTestRule<>(LocationLayerModesActivity.class);
@@ -65,105 +68,86 @@ public class LocationLayerPluginTest {
   }
 
   @Test
-  @Ignore
   public void locationSourceAdded() throws Exception {
     executeLocationLayerTest(new LocationLayerPluginAction.onPerformLocationLayerAction() {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController) {
         locationLayerPlugin.setLocationLayerEnabled(LocationLayerMode.TRACKING);
-        assertTrue(mapboxMap.getSource(LocationLayerConstants.LOCATION_SOURCE) != null);
+        assertTrue(mapboxMap.getSource(LOCATION_SOURCE) != null);
       }
     });
   }
 
   @Test
-  @Ignore
-  public void locationAccuracySourceAdded() throws Exception {
-    executeLocationLayerTest(new LocationLayerPluginAction.onPerformLocationLayerAction() {
-      @Override
-      public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
-                                        UiController uiController) {
-        locationLayerPlugin.setLocationLayerEnabled(LocationLayerMode.TRACKING);
-        assertTrue(mapboxMap.getSource(LocationLayerConstants.LOCATION_ACCURACY_SOURCE) != null);
-      }
-    });
-  }
-
-  @Test
-  @Ignore
   public void locationTrackingLayersAdded() throws Exception {
     executeLocationLayerTest(new LocationLayerPluginAction.onPerformLocationLayerAction() {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController) {
         locationLayerPlugin.setLocationLayerEnabled(LocationLayerMode.TRACKING);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_ACCURACY_LAYER) != null);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_BACKGROUND_LAYER) != null);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_LAYER) != null);
+        assertTrue(mapboxMap.getLayer(ACCURACY_LAYER) != null);
+        assertTrue(mapboxMap.getLayer(BACKGROUND_LAYER) != null);
+        assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER) != null);
       }
     });
   }
 
   @Test
-  @Ignore
   public void locationBearingLayersAdded() throws Exception {
     executeLocationLayerTest(new LocationLayerPluginAction.onPerformLocationLayerAction() {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController) {
         locationLayerPlugin.setLocationLayerEnabled(LocationLayerMode.COMPASS);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_ACCURACY_LAYER) != null);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_BACKGROUND_LAYER) != null);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_LAYER) != null);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_BEARING_LAYER) != null);
+        assertTrue(mapboxMap.getLayer(ACCURACY_LAYER) != null);
+        assertTrue(mapboxMap.getLayer(BACKGROUND_LAYER) != null);
+        assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER) != null);
+        assertTrue(mapboxMap.getLayer(BEARING_LAYER) != null);
       }
     });
   }
 
   @Test
-  @Ignore
   public void locationNavigationLayersAdded() throws Exception {
     executeLocationLayerTest(new LocationLayerPluginAction.onPerformLocationLayerAction() {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController) {
         locationLayerPlugin.setLocationLayerEnabled(LocationLayerMode.COMPASS);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_NAVIGATION_LAYER) != null);
+        assertTrue(mapboxMap.getLayer(NAVIGATION_LAYER) != null);
       }
     });
   }
 
   @Test
-  @Ignore
   public void locationLayerModeCorrectlySetToNone() throws Exception {
     executeLocationLayerTest(new LocationLayerPluginAction.onPerformLocationLayerAction() {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController) {
         locationLayerPlugin.setLocationLayerEnabled(LocationLayerMode.TRACKING);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_LAYER) != null);
+        assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER) != null);
         locationLayerPlugin.setLocationLayerEnabled(LocationLayerMode.NONE);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_LAYER).getVisibility().getValue()
+        assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER).getVisibility().getValue()
           .equals(Property.NONE));
       }
     });
   }
 
   @Test
-  @Ignore
   public void onMapChangeLocationLayerRedrawn() throws Exception {
     executeLocationLayerTest(new LocationLayerPluginAction.onPerformLocationLayerAction() {
       @Override
       public void onLocationLayerAction(LocationLayerPlugin locationLayerPlugin, MapboxMap mapboxMap,
                                         UiController uiController) {
         locationLayerPlugin.setLocationLayerEnabled(LocationLayerMode.TRACKING);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_LAYER) != null);
+        assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER) != null);
         mapboxMap.setStyleUrl(Style.SATELLITE);
         uiController.loopMainThreadForAtLeast(500);
         assertEquals(locationLayerPlugin.getLocationLayerMode(), LocationLayerMode.TRACKING);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_LAYER) != null);
-        assertTrue(mapboxMap.getLayer(LocationLayerConstants.LOCATION_LAYER).getVisibility().getValue()
+        assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER) != null);
+        assertTrue(mapboxMap.getLayer(FOREGROUND_LAYER).getVisibility().getValue()
           .equals(Property.VISIBLE));
       }
     });