+++ app/src/main/java/com/mapbox/mapboxsdk/plugins/testapp/activity/location/LocationLayerModesActivity.java
@@ -1,11 +1,12 @@
 package com.mapbox.mapboxsdk.plugins.testapp.activity.location;
 
-import android.arch.lifecycle.LifecycleRegistry;
-import android.arch.lifecycle.LifecycleRegistryOwner;
 import android.location.Location;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
+import android.widget.Toast;
 
 import com.mapbox.mapboxsdk.camera.CameraUpdateFactory;
 import com.mapbox.mapboxsdk.geometry.LatLng;
@@ -14,6 +15,7 @@ import com.mapbox.mapboxsdk.maps.MapboxMap;
 import com.mapbox.mapboxsdk.maps.OnMapReadyCallback;
 import com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerMode;
 import com.mapbox.mapboxsdk.plugins.locationlayer.LocationLayerPlugin;
+import com.mapbox.mapboxsdk.plugins.locationlayer.OnLocationLayerClickListener;
 import com.mapbox.mapboxsdk.plugins.testapp.R;
 import com.mapbox.services.android.location.LostLocationEngine;
 import com.mapbox.services.android.telemetry.location.LocationEngine;
@@ -25,22 +27,21 @@ import butterknife.ButterKnife;
 import butterknife.OnClick;
 
 public class LocationLayerModesActivity extends AppCompatActivity implements OnMapReadyCallback,
-  LifecycleRegistryOwner, LocationEngineListener {
+  LocationEngineListener, OnLocationLayerClickListener {
 
   @BindView(R.id.mapView)
   MapView mapView;
 
   private LocationLayerPlugin locationLayerPlugin;
-  private LifecycleRegistry lifecycleRegistry;
   private LocationEngine locationEngine;
   private MapboxMap mapboxMap;
+  private boolean customStyle;
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
     setContentView(R.layout.activity_location_layer_mode);
     ButterKnife.bind(this);
-    lifecycleRegistry = new LifecycleRegistry(this);
 
     mapView.onCreate(savedInstanceState);
     mapView.getMapAsync(this);
@@ -90,16 +91,33 @@ public class LocationLayerModesActivity extends AppCompatActivity implements OnM
     locationEngine.addLocationEngineListener(this);
     locationEngine.activate();
     locationLayerPlugin = new LocationLayerPlugin(mapView, mapboxMap, locationEngine);
+    locationLayerPlugin.setOnLocationClickListener(this);
     getLifecycle().addObserver(locationLayerPlugin);
+  }
+
+  @Override
+  public boolean onCreateOptionsMenu(Menu menu) {
+    getMenuInflater().inflate(R.menu.menu_location, menu);
+    return true;
+  }
+
+  @Override
+  public boolean onOptionsItemSelected(MenuItem item) {
+    if (locationLayerPlugin == null) {
+      return super.onOptionsItemSelected(item);
+    }
+
+    if (item.getItemId() == R.id.action_style_change) {
+      customStyle = !customStyle;
+      locationLayerPlugin.applyStyle(customStyle ? R.style.CustomLocationLayer : R.style.LocationLayer);
+      return true;
+    }
+
+    return super.onOptionsItemSelected(item);
   }
 
   public LocationLayerPlugin getLocationLayerPlugin() {
     return locationLayerPlugin;
-  }
-
-  @Override
-  public LifecycleRegistry getLifecycle() {
-    return lifecycleRegistry;
   }
 
   @Override
@@ -166,5 +184,10 @@ public class LocationLayerModesActivity extends AppCompatActivity implements OnM
   public void onLocationChanged(Location location) {
     mapboxMap.animateCamera(CameraUpdateFactory.newLatLngZoom(
       new LatLng(location.getLatitude(), location.getLongitude()), 16));
+  }
+
+  @Override
+  public void onLocationLayerClick() {
+    Toast.makeText(this, "OnLocationLayerClick", Toast.LENGTH_LONG).show();
   }
 }