+++ app/src/main/java/com/mapbox/mapboxsdk/plugins/testapp/activity/offline/OfflineDownloadActivity.java
@@ -6,6 +6,8 @@ import android.widget.ArrayAdapter;
 import android.widget.EditText;
 import android.widget.SeekBar;
 import android.widget.Spinner;
+import android.widget.TextView;
+import android.widget.Toast;
 
 import com.mapbox.mapboxsdk.constants.MapboxConstants;
 import com.mapbox.mapboxsdk.constants.Style;
@@ -45,6 +47,12 @@ public class OfflineDownloadActivity extends AppCompatActivity {
   @BindView(R.id.edittext_lon_west)
   EditText lonWestView;
 
+  @BindView(R.id.min_text_view)
+  TextView minTextView;
+
+  @BindView(R.id.max_text_view)
+  TextView maxTextView;
+
   @BindView(R.id.spinner_style_url)
   Spinner styleUrlView;
 
@@ -60,12 +68,14 @@ public class OfflineDownloadActivity extends AppCompatActivity {
     setContentView(R.layout.activity_offline_download);
     ButterKnife.bind(this);
     initUi();
+    initSeekbarListeners();
   }
 
   private void initUi() {
     initEditTexts();
     initSeekbars();
     initSpinner();
+    initZoomLevelTextviews();
   }
 
   private void initEditTexts() {
@@ -95,6 +105,47 @@ public class OfflineDownloadActivity extends AppCompatActivity {
     styleUrlView.setAdapter(spinnerArrayAdapter);
   }
 
+  private void initZoomLevelTextviews() {
+    maxTextView.setText(getString(R.string.max_zoom_textview, maxZoomView.getProgress()));
+    minTextView.setText(getString(R.string.min_zoom_textview, minZoomView.getProgress()));
+  }
+
+  private void initSeekbarListeners() {
+    maxZoomView.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
+      @Override
+      public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
+        maxTextView.setText(getString(R.string.max_zoom_textview, progress));
+      }
+
+      @Override
+      public void onStartTrackingTouch(SeekBar seekBar) {
+
+      }
+
+      @Override
+      public void onStopTrackingTouch(SeekBar seekBar) {
+
+      }
+    });
+
+    minZoomView.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
+      @Override
+      public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
+        minTextView.setText(getString(R.string.min_zoom_textview, progress));
+      }
+
+      @Override
+      public void onStartTrackingTouch(SeekBar seekBar) {
+
+      }
+
+      @Override
+      public void onStopTrackingTouch(SeekBar seekBar) {
+
+      }
+    });
+  }
+
   @OnClick(R.id.fab)
   public void onDownloadRegion() {
     // get data from UI
@@ -106,6 +157,11 @@ public class OfflineDownloadActivity extends AppCompatActivity {
     String styleUrl = (String) styleUrlView.getSelectedItem();
     float maxZoom = maxZoomView.getProgress();
     float minZoom = minZoomView.getProgress();
+
+    if (!validCoordinates(latitudeNorth, longitudeEast, latitudeSouth, longitudeWest)) {
+      Toast.makeText(this, "coordinates need to be in valid range", Toast.LENGTH_LONG).show();
+      return;
+    }
 
     // create offline definition from data
     OfflineTilePyramidRegionDefinition definition = new OfflineTilePyramidRegionDefinition(
@@ -131,5 +187,19 @@ public class OfflineDownloadActivity extends AppCompatActivity {
         .withMetadata(OfflineUtils.convertRegionName(regionName))
         .withNotificationOptions(notificationOptions)
     );
+  }
+
+  private boolean validCoordinates(double latitudeNorth, double longitudeEast, double latitudeSouth,
+                                   double longitudeWest) {
+    if (latitudeNorth < -90 || latitudeNorth > 90) {
+      return false;
+    } else if (longitudeEast < -180 || longitudeEast > 180) {
+      return false;
+    } else if (latitudeSouth < -90 || latitudeSouth > 90) {
+      return false;
+    } else if (longitudeWest < -180 || longitudeWest > 180) {
+      return false;
+    }
+    return true;
   }
 }