+++ plugin-places/src/main/java/com/mapbox/plugins/places/autocomplete/viewmodel/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Contains context aware view models for preparing and managing the data that place activities use.
+ *
+ * @since 0.1.0
+ */
+package com.mapbox.plugins.places.autocomplete.viewmodel;