+++ plugin-locationlayer/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Mapbox welcomes participation and contributions from everyone.
 
+### 0.3.0 - December 1, 2017
+- Change accuracy ring from fill layer to a circle layer [#186](https://github.com/mapbox/mapbox-plugins-android/pull/186)
+- Adds last known location API [#199](https://github.com/mapbox/mapbox-plugins-android/pull/199)
+- Adds location layer icon click listener [#198](https://github.com/mapbox/mapbox-plugins-android/pull/198)
+- Rework location layer abstraction [#196](https://github.com/mapbox/mapbox-plugins-android/pull/196)
+
 ### 0.2.0 - August 18, 2017
 - Updated Map and Mapbox Java Dependencies [#84](https://github.com/mapbox/mapbox-plugins-android/pull/84)
 - Navigation icon now uses runtime styling to scale at lower zoom levels [#84](https://github.com/mapbox/mapbox-plugins-android/pull/84)
@@ -9,3 +15,4 @@ Mapbox welcomes participation and contributions from everyone.
 
 ### 0.1.0 - July 17, 2017
 - Initial release as a standalone package.
+